chore(file_structure): Refactor components into folders
