+++ amplify/backend/function/AdminQueries68290a65/AdminQueries68290a65-cloudformation-template.json
@@ -67,8 +67,8 @@
 				"Runtime": "nodejs10.x",
 				"Timeout": "25",
 				"Code": {
-					"S3Bucket": "amplify-robobot-dev-191612-deployment",
-					"S3Key": "amplify-builds/AdminQueries68290a65-4e445632645864743167-build.zip"
+					"S3Bucket": "amplify-robobot-devtest-173109-deployment",
+					"S3Key": "amplify-builds/AdminQueries68290a65-48616a6464313843664a-build.zip"
 				}
 			}
 		},
