+++ src/components/game_modal/index.css
