+++ src/components/horizontal_split_layout/index.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState, useContext, useRef } from "react";
 
 import { Row, Col } from "antd";
 
-import { useWindowSize } from "../hooks/useWindowSize";
+import { useWindowSize } from "hooks/useWindowSize";
 
 // Dynamically allocates remaining vertical screen-space to the dependent component
 // dependent = "top" or dependent = "bottom"
