+++ src/components/split_container/index.js
